refactor(middlewares): express endpoints into separated files
