+++ middlewares/ping.js
@@ -1,0 +1,6 @@
+//
+// Endpoint to check the health of the application.
+//
+export default () => (req, res) => {
+  return res.end(JSON.stringify({ status: 'ok' }))
+}
